// src/atomic/components/dropdown/dropdown.enums.ts
export enum DDSize {
  XLarge = 'h2--xl',
  Large = 'h2--l',
  Medium = 'h2--m',
  Small = 'h2--s',
  XSmall = 'h2--xs',
  XXSmall = 'h2--xxs',
}

export enum DDTemplate {
  Text = 'textTemplate',
  LineThickness = 'lineThicknessTemplate',
  LineStyle = 'lineStyleTemplate',
  MarkerStyle = 'markerStyleTemplate',
  Command = 'commandTemplate',
  ColorGradient = 'colorGradientTemplate'
}

export enum DDTexAlign {
  Left = 'u-text-left',
  Right = 'u-text-right',
  Center = 'u-text-center'
}